+++ enpt/options/options_schema.py
@@ -41,6 +41,7 @@ enpt_schema_input = dict(
             atmospheric_correction=dict(
                 type='dict', required=False,
                 schema=dict(
+                    sicor_cache_dir=dict(type='string', required=False),
                     auto_download_ecmwf=dict(type='boolean', required=False),
                     enable_cloud_screening=dict(type='boolean', required=False),
                 )),
@@ -55,6 +56,9 @@ enpt_schema_input = dict(
                 type='dict', required=False,
                 schema=dict(
                     run_processor=dict(type='boolean', required=False),
+                    algorithm=dict(type='string', required=False, allowed=['spectral', 'spatial']),
+                    interpolation_method=dict(type='string', required=False,
+                                              allowed=['linear', 'bilinear', 'cubic', 'spline']),
                 )),
 
             orthorectification=dict(
@@ -82,6 +86,7 @@ parameter_mapping = dict(
     # processors > toa_ref
     path_earthSunDist=('processors', 'toa_ref', 'path_earthSunDist'),
     path_solar_irr=('processors', 'toa_ref', 'path_solar_irr'),
+    scale_factor_toa_ref=('processors', 'toa_ref', 'scale_factor_toa_ref'),
 
     # processors > geometry
     enable_keystone_correction=('processors', 'geometry', 'enable_keystone_correction'),
@@ -89,14 +94,18 @@ parameter_mapping = dict(
     path_reference_image=('processors', 'geometry', 'path_reference_image'),
 
     # processors > atmospheric_correction
+    sicor_cache_dir=('processors', 'atmospheric_correction', 'sicor_cache_dir'),
     auto_download_ecmwf=('processors', 'atmospheric_correction', 'auto_download_ecmwf'),
     enable_cloud_screening=('processors', 'atmospheric_correction', 'enable_cloud_screening'),
+    scale_factor_boa_ref=('processors', 'atmospheric_correction', 'scale_factor_boa_ref'),
 
     # processors > smile
     run_smile_P=('processors', 'smile', 'run_processor'),
 
     # processors > dead_pixel
     run_deadpix_P=('processors', 'dead_pixel', 'run_processor'),
+    deadpix_P_algorithm=('processors', 'dead_pixel', 'algorithm'),
+    deadpix_P_interp=('processors', 'dead_pixel', 'interpolation_method'),
 
     # processors > orthorectification
     ortho_resampAlg=('processors', 'orthorectification', 'resamp_alg'),