+++ enpt/processors/dead_pixel_correction/__init__.py
@@ -1,2 +1,6 @@
 # -*- coding: utf-8 -*-
 """EnPT 'dead pixel correction' module for detecting and correcting dead pixels."""
+
+from .dead_pixel_correction import Dead_Pixel_Corrector
+
+__all__ = ['Dead_Pixel_Corrector']