+++ HISTORY.rst
@@ -2,13 +2,27 @@
 History
 =======
 
-0.x.x (2018-05-16)
-------------------
+0.x.x (coming soon)
+-------------------
 New features:
 
 * Use a new header file
 * Add a second image at the bottom of the main image (if possible), select a possible number of line
 * Save extended image
+
+
+0.4.0 (2018-06-01)
+------------------
+New features:
+* Implemented dead pixel corrector
+* Implemented SICOR atmospheric correction
+
+
+0.3.0 (??)
+----------
+
+New features:
+* TBD
 
 
 0.2.0 (2017-08-24)